moving social links static array
